ToDo 5: Fix the error ToDo 6: Fix the constructor of Student class

+++ src/main/java/com/mycompany/csc325_oop_designreview_lab/MainClass.java
@@ -11,9 +11,7 @@ package com.mycompany.csc325_oop_designreview_lab;
 public class MainClass {
 	
 	public static void main(String[] args) {
-		// ToDo 5: Fix the error
-		
-		// ToDo 6: Fix the constructor of Student class
+
                 
                 // Todo 7: Create two classes for Freshman and Senior 
                 
@@ -22,8 +20,8 @@ public class MainClass {
 		// ToDo 9: Add a toString method for Freshman class
 		
                 // ToDo 10: Add a toString method for Senior class
-                
-		Freshman std1= new Student("James", 20, 12); // name, age, credits
+             
+		Freshman std1= new Student("James", (short)20, 12); // name, age, credits
                 
                 Senior std2 = new Student("John", 30, 90);
